split out the index module into a separate file and do some small refactorings
