+++ .gitignore
@@ -1,1 +1,4 @@
 /target
+
+**/db_file.csv
+
